Replace deprecated QMessageBox::information()/warning()
With Qt6, those formats of QMessageBox::information() that take
non-standard label texts have been deprecated. The intention is
to always use standard texts for labels.

QMessageBox::warning():

The numeric arguments were triggering an obsolete form of the
warning dialog. Instead, omit the last "0" so it is now taken
as a supported version with overloaded button names consisting
of just one button.