Enable PTT if there is an RTS or DTR PTT configured
Using hamlib's capability reporting only reports capabilities
that are available straight through the CAT protocol of the
target rig. PTTs controlled by RTS or DTR are out-of-band
with respect to the CAT protocol, so they do not appear there.
Nevertheless, if the operator configured such a PTT type, we
assume the rig PTT can be activated which can then still be
done using rig_set_ptt().