Implement the option "no PTT" in CAT configuration
If the respective CAT configuration does not support a PTT
(e.g. since the rig's CAT protocol does not implement it,
and there is no DTR- or RTS-controlled hardware either),
it is possible to turn of PTT entirely yet still enable
CAT for frequency and mode control. (PTT could then be
performced e.g. via VOX.)